exo-todolist logpage added

+++ exo-todoList/app/remove.php
@@ -1,5 +1,10 @@
 <?php 
+session_start();
 
+if (!isset($_SESSION['id']) || !isset($_SESSION['user_name'])) {
+    echo "error";
+    exit();
+}
 if(isset($_POST['id'])){
     require '../db_conn.php';
     
@@ -9,8 +14,8 @@ if(isset($_POST['id'])){
     if(empty($id)){
         echo 0;
     }else {
-        $stmt = $conn->prepare("DELETE FROM todos WHERE id=? ");
-        $res = $stmt->execute([$id]);
+        $stmt = $conn->prepare("DELETE FROM todos WHERE id=?  and userId=?");
+        $res = $stmt->execute([$id, $_SESSION['id']]);
 
         if($res){
             echo 1;
